+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Added
 - it's now possible to compare Values using `operator<`
 - `reverseList` (added to the FFI) by @rinz13r
+- a warning will now pop up during compilation if code appears to be ill-formed (eg number/string/list right after a `(`)
 
 ### Changed
 - the internal API used to compare values has been updated to be much shorter